move assert*Response magic to a new class
This allows us to break the tests to smaller modules to improve testing time.
